Add write completion channels.
Write completion is an important event for correctly sequencing memory accesses. Before this change, abstract RAMs had a "abstract_write_resp" (a channel with type `()` that was a pure write completion channel) and 1RW RAMs had a "resp" channel. The 1RW "resp" channel has data- in the event of a write, you are supposed to recv() with predicate set to false (this avoids latching garbage). Upcoming 1R1W implementations also need a notion of write completion, but unlike 1RW it would be incorrect the use the read response channel as a proxy for write completion as reads and writes should be schedulable independently.

This deals with these issues by creating a "write completion" channel common between all the RAMs. Conceivably, other variants could use a different write completion channel (e.g. variable latency memory w/ IDs), but for the immediate future it seems like most memories will be able to share this kind of write completion channel.

The write completion channel is removed in codegen and its valid signal is set to literal 1.

Working towards google/xls#861.

PiperOrigin-RevId: 518117163