Return stderr to buffered IO.
Under heavy load NDELAY messes up the propagation of close events on the nonblocking pipe and that causes zombie processes as the code "waits forever" in ReadFileDescriptors and never ends up reading the exit status that comes back in waitpid().

Without this change, run-clang-tidy-cached.sh generates lots of zombies. With this change it does not produce any.

PiperOrigin-RevId: 518248076