[DSLX] Give a good type inference error when channel is used in wrong position.
Previously we were reaching in to AST nodes to determine channel information,
but it's more uniform to just call the type inference deduction machinery and
see that the resulting type is a (half-duplex) Channel type as we expect in
that position.

In the process discovered that the sender/receiver for channel declarations
could be arbitrarily reversed! Now it is canonically:

```
  let (sender, receiver) = chan<u32>;
```

This CL fixes a bunch of fallout from that.

PiperOrigin-RevId: 519196041