[DSLX] Deduce type AST nodes to be of type "MetaType" (i.e. a type-type).
Before, as a bit of a kluge, we'd just say that types yielded their own type;
but really only expressions should yield value of that type.

This lets us distinguish in the type system when type annotations are
inappropriately used in expression positions.

Note: we use the term "metatype" for the moment instead of "kind" since our
only representation is a trivial kind ("the type of some other type" which we
can't ask any meaningful properties about in any generic sense).  With more
generic support we'll want to switch to kind structures with meaningful
attributes.

Fixes google/xls#840 (the more-innocuous-looking than it is :-)

PiperOrigin-RevId: 521604390