Schedule sends/recvs in different logical cycles in proc clumping
Previously, there was nothing stopping the proc clumping pass from
scheduling sends/receives on the same channel in the same logical
cycle, thereby preventing the mutual exclusion scheduling pass from
merging them together. This change separately topologically sorts
the sends and receives on a given channel, and then passes in
constraints that those ops must be scheduled in that order.

The benefit of using a topological order is that these constraints
cannot contradict an existing causality constraint. However, it is
possible to do better.

PiperOrigin-RevId: 527299567