[DSLX] Break out abstract interface for the RunComparator.
This makes a virtual seam that avoids the DSLX run routines coupling directly
to e.g. the JIT facilities accessed by the RunComparison object.

This will allow us to support configurations where the JIT is not present.

To accomplish this the change also needed to update quickcheck to be capable of
running with the IR interpreter as well as the IR JIT.

Helpful towards google/xls#962

PiperOrigin-RevId: 532974312